fix(signal): extract text-style parser to satisfy LOC ratchet
Move parseContainerTextStyleEntry into a sibling module so the oversized
client-container.ts does not grow while keeping strict integer parsing.
