+++ extensions/signal/src/client-container-text-style.ts
@@ -1,0 +1,18 @@
+// Parse Signal container text-style span entries.
+import { parseStrictNonNegativeInteger } from "openclaw/plugin-sdk/number-runtime";
+
+export function parseContainerTextStyleEntry(
+  raw: string,
+): { start: number; length: number; style: string } | undefined {
+  const [startRaw, lengthRaw, style] = raw.split(":");
+  if (startRaw === undefined || lengthRaw === undefined || style === undefined) {
+    return undefined;
+  }
+  // Match quote-timestamp: reject hex/exponent so style spans cannot silently shift.
+  const start = parseStrictNonNegativeInteger(startRaw);
+  const length = parseStrictNonNegativeInteger(lengthRaw);
+  if (start === undefined || length === undefined) {
+    return undefined;
+  }
+  return { start, length, style };
+}
